requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
colorama==0.3.9
dj-database-url==0.4.2
Django==1.9.6
django-celery==3.2.1
django-kombu==0.9.4
django-mail-templated==2.6.5
gunicorn==19.7.1
honcho==1.0.1
kombu==3.0.37
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
pytz==2017.2
